(fix) change the block size to be 5
Prior to this commit, the chunk size was 10, which meant 20 entries in
the block. The limit is 10, so this commit changes the chunk size to 5
since there are two elements in each chunk. This should fix an issue
where slack will not post the message if there are more than 5 instances
in a single state.