(maint) Add label validation and reporting
Prior to this commit, there was little label validation which would
cause unhelpful execptions to be shown in the slack message. This commit
adds validation logic to help clarify issues when instances are
mislabeled. It also notifies slack when instances are going to expire
and adds a simple verbose log level.